# Conflicts:
#	src/main/java/dev/clownsinformatics/tiendajava/rest/users/dto/UserInfoResponse.java

+++ src/main/java/dev/clownsinformatics/tiendajava/rest/users/controllers/UsersRestController.java
@@ -8,10 +8,16 @@ import dev.clownsinformatics.tiendajava.rest.orders.service.OrderService;
 import dev.clownsinformatics.tiendajava.rest.users.dto.UserInfoResponse;
 import dev.clownsinformatics.tiendajava.rest.users.dto.UserRequest;
 import dev.clownsinformatics.tiendajava.rest.users.dto.UserResponse;
+import dev.clownsinformatics.tiendajava.rest.users.exceptions.UnauthorizedUser;
 import dev.clownsinformatics.tiendajava.rest.users.models.User;
 import dev.clownsinformatics.tiendajava.rest.users.services.UsersService;
 import dev.clownsinformatics.tiendajava.utils.pagination.PageResponse;
 import dev.clownsinformatics.tiendajava.utils.pagination.PaginationLinksUtils;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.Parameters;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
@@ -64,6 +70,22 @@ public class UsersRestController {
      * @param request   petición
      * @return Respuesta con la página de usuarios
      */
+
+    @Operation(summary = "Find all users")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "username", required = false, description = "Username"),
+            @Parameter(name = "email", required = false, description = "Email"),
+            @Parameter(name = "isDeleted", required = false, description = "Is deleted"),
+            @Parameter(name = "page", required = false, description = "Page"),
+            @Parameter(name = "size", required = false, description = "Size"),
+            @Parameter(name = "sortBy", required = false, description = "Sort by"),
+            @Parameter(name = "direction", required = false, description = "Direction"),
+            @Parameter(name = "request", required = true, description = "Request")
+    })
     @GetMapping
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<PageResponse<UserResponse>> findAll(
@@ -78,9 +100,7 @@ public class UsersRestController {
     ) {
         log.info("findAll: username: {}, email: {}, isDeleted: {}, page: {}, size: {}, sortBy: {}, direction: {}",
                 username, email, isDeleted, page, size, sortBy, direction);
-        // Creamos el objeto de ordenación
         Sort sort = direction.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
-        // Creamos cómo va a ser la paginación
         UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromHttpUrl(request.getRequestURL().toString());
         Page<UserResponse> pageResult = usersService.findAll(username, email, isDeleted, PageRequest.of(page, size, sort));
         return ResponseEntity.ok()
@@ -95,8 +115,16 @@ public class UsersRestController {
      * @return Usuario si existe
      * @throws dev.clownsinformatics.tiendajava.rest.users.exceptions.UserNotFound si no existe el usuario (404)
      */
+    @Operation(summary = "Find user by id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "id", required = true, description = "User id")
+    })
     @GetMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')") // Solo los admin pueden acceder
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<UserInfoResponse> findById(@PathVariable Long id) {
         log.info("findById: id: {}", id);
         return ResponseEntity.ok(usersService.findById(id));
@@ -107,11 +135,20 @@ public class UsersRestController {
      *
      * @param userRequest usuario a crear
      * @return Usuario creado
-     * @throws dev.clownsinformatics.tiendajava.rest.users.exceptions.UserNameOrEmailExists               si el nombre de usuario o el email ya existen
-     * @throws HttpClientErrorException.BadRequest si hay algún error de validación
+     * @throws dev.clownsinformatics.tiendajava.rest.users.exceptions.UserNameOrEmailExists si el nombre de usuario o el email ya existen
+     * @throws HttpClientErrorException.BadRequest                                          si hay algún error de validación
      */
+    @Operation(summary = "Create user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Created"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "userRequest", required = true, description = "User data")
+    })
     @PostMapping
-    @PreAuthorize("hasRole('ADMIN')") // Solo los admin pueden acceder
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<UserResponse> createUser(@Valid @RequestBody UserRequest userRequest) {
         log.info("save: userRequest: {}", userRequest);
         return ResponseEntity.status(HttpStatus.CREATED).body(usersService.save(userRequest));
@@ -127,8 +164,19 @@ public class UsersRestController {
      * @throws HttpClientErrorException.BadRequest                                          si hay algún error de validación (400)
      * @throws dev.clownsinformatics.tiendajava.rest.users.exceptions.UserNameOrEmailExists si el nombre de usuario o el email ya existen (400)
      */
+    @Operation(summary = "Update user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "404", description = "Not Found"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "id", required = true, description = "User id"),
+            @Parameter(name = "userRequest", required = true, description = "User data")
+    })
     @PutMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')") // Solo los admin pueden acceder
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @Valid @RequestBody UserRequest userRequest) {
         log.info("update: id: {}, userRequest: {}", id, userRequest);
         return ResponseEntity.ok(usersService.update(id, userRequest));
@@ -141,8 +189,18 @@ public class UsersRestController {
      * @return Respuesta vacía
      * @throws dev.clownsinformatics.tiendajava.rest.users.exceptions.UserNotFound si no existe el usuario (404)
      */
+    @Operation(summary = "Delete user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "No Content"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "404", description = "Not Found"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "id", required = true, description = "User id")
+    })
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')") // Solo los admin pueden acceder
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
         log.info("delete: id: {}", id);
         usersService.deleteById(id);
@@ -155,6 +213,14 @@ public class UsersRestController {
      * @param user usuario autenticado
      * @return Datos del usuario
      */
+    @Operation(summary = "Get current user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data")
+    })
     @GetMapping("/me/profile")
     public ResponseEntity<UserInfoResponse> me(@AuthenticationPrincipal User user) {
         log.info("Obteniendo usuario");
@@ -169,8 +235,17 @@ public class UsersRestController {
      * @return Usuario actualizado
      * @throws HttpClientErrorException.BadRequest si hay algún error de validación (400)
      */
+    @Operation(summary = "Update current user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "userRequest", required = true, description = "User data")
+    })
     @PutMapping("/me/profile")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<UserResponse> updateMe(@AuthenticationPrincipal User user, @Valid @RequestBody UserRequest userRequest) {
         log.info("updateMe: user: {}, userRequest: {}", user, userRequest);
         return ResponseEntity.ok(usersService.update(user.getId(), userRequest));
@@ -182,8 +257,16 @@ public class UsersRestController {
      * @param user usuario autenticado
      * @return Respuesta vacía
      */
+    @Operation(summary = "Delete current user")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "No Content"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data")
+    })
     @DeleteMapping("/me/profile")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<Void> deleteMe(@AuthenticationPrincipal User user) {
         log.info("deleteMe: user: {}", user);
         usersService.deleteById(user.getId());
@@ -200,9 +283,20 @@ public class UsersRestController {
      * @param direction dirección de ordenación
      * @return Respuesta con la página de pedidos
      */
-
+    @Operation(summary = "Get current user orders")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "page", required = false, description = "Page"),
+            @Parameter(name = "size", required = false, description = "Size"),
+            @Parameter(name = "sortBy", required = false, description = "Sort by"),
+            @Parameter(name = "direction", required = false, description = "Direction")
+    })
     @GetMapping("/me/pedidos")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<PageResponse<OrderResponseDto>> getPedidosByUsuario(
             @AuthenticationPrincipal User user,
             @RequestParam(defaultValue = "0") int page,
@@ -217,38 +311,57 @@ public class UsersRestController {
     }
 
 
-
-/**
+    /**
      * Obtiene un pedido del usuario actual
      *
      * @param user     usuario autenticado
      * @param idPedido id del pedido
      * @return Pedido
      */
-
+    @Operation(summary = "Get current user order")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "idPedido", required = true, description = "Order id")
+    })
     @GetMapping("/me/pedidos/{id}")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<OrderResponseDto> getPedido(
             @AuthenticationPrincipal User user,
             @PathVariable("id") ObjectId idPedido
     ) {
         log.info("Obteniendo pedido con id: " + idPedido);
+        if (!orderService.findById(idPedido).idUser().equals(user.getId())) {
+            throw new UnauthorizedUser("The user is not authorized to get this order");
+        }
         return ResponseEntity.ok(orderService.findById(idPedido));
     }
 
 
-
-/**
+    /**
      * Crea un pedido para el usuario actual
      *
      * @param user   usuario autenticado
      * @param pedido pedido a crear
      * @return Pedido creado
      * @throws HttpClientErrorException.BadRequest si hay algún error de validación (400)
+     * @throws HttpClientErrorException.Unauthorized si el usuario no está autorizado (403)
      */
-
+    @Operation(summary = "Create current user order")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Created"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "pedido", required = true, description = "Order data")
+    })
     @PostMapping("/me/pedidos")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<OrderResponseDto> savePedido(
             @AuthenticationPrincipal User user,
             @Valid @RequestBody OrderCreateDto pedido
@@ -258,7 +371,7 @@ public class UsersRestController {
     }
 
 
-/**
+    /**
      * Actualiza un pedido del usuario actual
      *
      * @param user     usuario autenticado
@@ -266,39 +379,68 @@ public class UsersRestController {
      * @param pedido   pedido a actualizar
      * @return Pedido actualizado
      * @throws HttpClientErrorException.BadRequest si hay algún error de validación (400)
+     * @throws HttpClientErrorException.Unauthorized si el usuario no está autorizado (403)
      */
-
+    @Operation(summary = "Update current user order")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "OK"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "idPedido", required = true, description = "Order id"),
+            @Parameter(name = "pedido", required = true, description = "Order data")
+    })
     @PutMapping("/me/pedidos/{id}")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<OrderResponseDto> updatePedido(
             @AuthenticationPrincipal User user,
             @PathVariable("id") ObjectId idPedido,
             @Valid @RequestBody OrderUpdateDto pedido) {
         log.info("Actualizando pedido con id: " + idPedido);
+        OrderResponseDto order = orderService.findById(idPedido);
+        if (!order.idUser().equals(user.getId())) {
+            throw new UnauthorizedUser("The user is not authorized to update this order");
+        }
         return ResponseEntity.ok(orderService.update(idPedido, pedido));
     }
 
 
-/**
+    /**
      * Borra un pedido del usuario actual
      *
      * @param user     usuario autenticado
      * @param idPedido id del pedido
      * @return Pedido borrado
      * @throws HttpClientErrorException.BadRequest si hay algún error de validación (400)
+     * @throws HttpClientErrorException.Unauthorized si el usuario no está autorizado (403)
      */
-
+    @Operation(summary = "Delete current user order")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "No Content"),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "404", description = "Not Found"),
+            @ApiResponse(responseCode = "403", description = "Forbidden")
+    })
+    @Parameters({
+            @Parameter(name = "user", required = true, description = "User data"),
+            @Parameter(name = "idPedido", required = true, description = "Order id")
+    })
     @DeleteMapping("/me/pedidos/{id}")
-    @PreAuthorize("hasRole('USER')") // Solo los usuarios pueden acceder
+    @PreAuthorize("hasRole('USER')")
     public ResponseEntity<Void> deletePedido(
             @AuthenticationPrincipal User user,
             @PathVariable("id") ObjectId idPedido
     ) {
         log.info("Borrando pedido con id: " + idPedido);
+        OrderResponseDto order = orderService.findById(idPedido);
+        if (!order.idUser().equals(user.getId())) {
+            throw new UnauthorizedUser("The user is not authorized to delete this order");
+        }
         orderService.delete(idPedido);
         return ResponseEntity.noContent().build();
     }
-
 
 
     /**
